requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=1.7
streamlit>=1.30
plotly>=5.0
shap>=0.44
joblib>=1.3
python-pptx>=0.6
reportlab>=4.0
matplotlib>=3.7
seaborn>=0.12
optuna>=3.0
kaleido>=0.2
